github-releases: extract digest.ts
